Rename and implement gitlab project service
